requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
pyodbc==4.0.32
SQLAlchemy-Utils==0.37.8
Flask-WTF==0.15.1   # Form işlemleri için
email-validator==1.1.3  # Email doğrulama için
Flask-JWT-Extended==4.4.4
python-dateutil==2.8.2
google-generativeai==0.3.2  # Gemini API için